backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
openai-whisper @ git+https://github.com/openai/whisper.git
